add Draw.java

+++ src/math/components/Point2.java
@@ -1,15 +1,15 @@
 package math.components;
 
 public class Point2 {
-    public double x, y;
+    public float x, y;
     String name;
-    public Point2(double x, double y, String name) {
+    public Point2(float x, float y, String name) {
         this.x = x;
         this.y = y;
         this.name = name;
     }
 
-    public Point2(double x, double y) {this(x, y, null);}
+    public Point2(float x, float y) {this(x, y, null);}
 
     public Point2 add(Point2 point) {
         return new Point2(this.x + point.x, this.y + point.y, point.name);
